# Optimise `ActionDispatch::Http::URL.build_host_url` when protocol in host