@joshuay03 2025-09-09
> Could the three branches be combined with something like this?

This line `protocol = protocol_from_host || normalize_protocol(protocol).dup` prioritises host-extracted protocols over explicit user options, breaking the expected behavior where user-provided `protocol` options should take precedence (`protocol ||= match[1] unless protocol == false`). We could replace it with this:
```ruby
  if protocol
    protocol = normalize_protocol(protocol).dup
  else
    protocol = protocol_from_host || normalize_protocol(nil)
  end
```

but I think the current approach is more idiomatic.