@skipkayhil 2025-09-09
I added a test on main that fails with the current code (addc45c5b4242a899daf4bea02038a8a29c18556), and added a commit to fix it. Let me know if it looks good and I'll squash + merge 👍