@joshuay03 2025-09-10
> I added a test on main that fails with the current code ([addc45c](https://github.com/rails/rails/commit/addc45c5b4242a899daf4bea02038a8a29c18556)), and added a commit to fix it. Let me know if it looks good and I'll squash + merge 👍

LGTM, thank you! Please feel free to add yourself to the changelog entry when squashing! :shipit: